refactor: move rejectSubmission to store
