+++ stores/moderationSubmissionsStore.ts
@@ -102,6 +102,11 @@ export const useModerationSubmissionsStore = defineStore(
         }
 
         async function approveSubmission() {
+            if (!selectedSubmissionId.value) {
+                console.error('Unable to approve submission. Submission ID is required.')
+                return
+            }
+
             const approveSubmissionInput: MutationUpdateSubmissionArgs = {
                 id: selectedSubmissionId.value,
                 input: {
@@ -109,6 +114,32 @@ export const useModerationSubmissionsStore = defineStore(
                 }
             }
             return updateSubmission(approveSubmissionInput)
+        }
+
+        async function rejectSubmission() {
+            if (!selectedSubmissionId.value) {
+                console.error('Unable to reject submission. Submission ID is required.')
+                return
+            }
+
+            const facilityInputVariables = {
+                updateSubmissionId: selectedSubmissionId.value,
+                input: {
+                    isRejected: true
+                }
+            }
+
+            try {
+                await graphQLClientRequestWithRetry(
+                    gqlClient.request.bind(gqlClient),
+                    rejectFacilitySubmissionGqlMutation,
+                    facilityInputVariables
+                )
+                setDidMutationFail(false)
+            } catch (error) {
+                console.error('Failed to reject submission:', error)
+                setDidMutationFail(true)
+            }
         }
 
         return { getSubmissions,
@@ -131,7 +162,8 @@ export const useModerationSubmissionsStore = defineStore(
             selectedModerationListViewTabChosen,
             setSelectedModerationListViewTabChosen,
             updateSubmission,
-            approveSubmission }
+            approveSubmission,
+            rejectSubmission }
     }
 )
 
@@ -238,5 +270,12 @@ mutation Mutation($id: ID!, $input: UpdateSubmissionInput!) {
       mapLatitude
       mapLongitude
     }
+  }
+}`
+
+const rejectFacilitySubmissionGqlMutation = gql`
+mutation Mutation($updateSubmissionId: ID!, $input: UpdateSubmissionInput!) {
+  updateSubmission(id: $updateSubmissionId, input: $input) {
+    isRejected
   }
 }`
